requirements.txt
pycryptodome
pyfiglet
PyInputPlus
python-secrets
stdiomask